ensure ChangeProfilePicture not calls DeleteFile if file does not exists and savePicture throws

+++ src/domain/use-cases/change-profile-picture.ts
@@ -20,7 +20,8 @@ export const setupChangeProfilePicture: SetupChangeProfilePicture = (fileStorage
   try {
     await userProfilePictureRepo.savePicture(userProfile)
   } catch {
-    await fileStorage.delete({ key })
+    if (input.file !== undefined) await fileStorage.delete({ key })
+    throw new Error()
   }
   return userProfile
 }
